move report generation to a new class
